Remove redundant cast.
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/work/src@39592 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ idps/src/com/untangle/node/idps/IdpsLogEvent.java
@@ -47,9 +47,9 @@ public class IdpsLogEvent extends LogEvent
         pstmt.setLong(++i, idpsEvent.getGeneratorId() );
         pstmt.setLong(++i, idpsEvent.getClassificationId() );
         pstmt.setObject(++i, idpsEvent.getIpSource().getHostAddress(), java.sql.Types.OTHER);
-        pstmt.setInt(++i, (int) ( idpsEvent.getSportItype() & 0xffff ) );
+        pstmt.setInt(++i, ( idpsEvent.getSportItype() & 0xffff ) );
         pstmt.setObject(++i, idpsEvent.getIpDestination().getHostAddress(), java.sql.Types.OTHER);
-        pstmt.setInt(++i, (int) ( idpsEvent.getDportIcode() & 0xffff ) );
+        pstmt.setInt(++i, ( idpsEvent.getDportIcode() & 0xffff ) );
         pstmt.setInt(++i, idpsEvent.getProtocol() );
         pstmt.setBoolean(++i, ( idpsEvent.getBlocked() == 1 ) ? true : false  );
